# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API server for ShipStation application",
        "title": "ShipStation RestAPI",
        "contact": {},
        "version": "1.0"
    },
    "host": "http://localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/ship": {
            "get": {
                "description": "Возращает список всех активных судов",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Судно"
                ],
                "summary": "Получение списка судов",
                "parameters": [
                    {
                        "type": "string",
                        "format": "email",
                        "description": "Название судна",
                        "name": "shipName",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список судов",
                        "schema": {
                            "$ref": "#/definitions/model.GetShips"
                        }
                    },
                    "400": {
                        "description": "Обработанная ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Создает новое судно с предоставленными данными",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Судно"
                ],
                "summary": "Создание нового судна",
                "parameters": [
                    {
                        "type": "string",
                        "format": "email",
                        "description": "Название судна",
                        "name": "shipName",
                        "in": "query"
                    },
                    {
                        "description": "Пользовательский объект в формате JSON",
                        "name": "ship",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ShipChange"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список судов",
                        "schema": {
                            "$ref": "#/definitions/model.GetShips"
                        }
                    },
                    "400": {
                        "description": "Обработанная ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "У пользователя нет прав для этого запроса",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/ship/{shipID}": {
            "get": {
                "description": "Возвращает информацию о суднe по его ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Судно"
                ],
                "summary": "Получение судна по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID судна",
                        "name": "shipID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о суднe",
                        "schema": {
                            "$ref": "#/definitions/model.Ship"
                        }
                    },
                    "400": {
                        "description": "Обработанная ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Обновляет информацию о суднe по его ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Судно"
                ],
                "summary": "Обновление информации о суднe",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID судна",
                        "name": "shipID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о суднe",
                        "schema": {
                            "$ref": "#/definitions/model.Ship"
                        }
                    },
                    "400": {
                        "description": "Обработанная ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "У пользователя нет прав для этого запроса",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Удаляет судно по его ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Судно"
                ],
                "summary": "Удаление судна",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID судна",
                        "name": "shipID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "email",
                        "description": "Название судна",
                        "name": "shipName",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список судов",
                        "schema": {
                            "$ref": "#/definitions/model.GetShips"
                        }
                    },
                    "400": {
                        "description": "Обработанная ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "У пользователя нет прав для этого запроса",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/ship/{shipID}/image": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Добавляет изображение к судноу по его ID",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Судно"
                ],
                "summary": "Добавление изображения к судноу",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID судна",
                        "name": "shipID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Изображение судна",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о суднe с изображением",
                        "schema": {
                            "$ref": "#/definitions/model.Ship"
                        }
                    },
                    "400": {
                        "description": "Обработанная ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "У пользователя нет прав для этого запроса",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/ship/{shipID}/request": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Добавляет судно к доставке по его ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Судно"
                ],
                "summary": "Добавление судна к доставке",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID судна",
                        "name": "shipID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "email",
                        "description": "Название судна",
                        "name": "shipName",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список судов",
                        "schema": {
                            "$ref": "#/definitions/model.GetShips"
                        }
                    },
                    "400": {
                        "description": "Обработанная ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Удаляет судно из доставки по его ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Судно"
                ],
                "summary": "Удаление судна из доставки",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID судна",
                        "name": "shipID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "email",
                        "description": "Название судна",
                        "name": "shipName",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список судов",
                        "schema": {
                            "$ref": "#/definitions/model.GetShips"
                        }
                    },
                    "400": {
                        "description": "Обработанная ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "model.GetShips": {
            "type": "object",
            "properties": {
                "request_id": {
                    "type": "integer"
                },
                "ships": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Ship"
                    }
                }
            }
        },
...
        "model.Ship": {
            "type": "object",
            "properties": {
                "cargo_capacity": {
                    "type": "number"
                },
                "classification": {
                    "type": "string"
                },
                "crew_capacity": {
                    "type": "integer"
                },
                "flag": {
                    "type": "string"
                },
                "max_depth": {
                    "type": "number"
                },
                "max_length": {
                    "type": "number"
                },
                "passenger_capacity": {
                    "type": "integer"
                },
                "photo": {
                    "type": "string"
                },
                "ship_id": {
                    "type": "integer"
                },
                "ship_name": {
                    "type": "string"
                },
                "ship_status": {
                    "type": "string"
                },
                "ship_type": {
                    "type": "string"
                },
                "year_built": {
                    "type": "integer"
                }
            }
        },
        "model.ShipChange": {
            "type": "object",
            "properties": {
                "cargo_capacity": {
                    "type": "number"
                },
                "classification": {
                    "type": "string"
                },
                "crew_capacity": {
                    "type": "integer"
                },
                "flag": {
                    "type": "string"
                },
                "max_depth": {
                    "type": "number"
                },
                "max_length": {
                    "type": "number"
                },
                "passenger_capacity": {
                    "type": "integer"
                },
                "ship_name": {
                    "type": "string"
                },
                "ship_type": {
                    "type": "string"
                },
                "year_built": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}